% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.R
\name{affinity_matrix}
\alias{affinity_matrix}
\title{Extract affinity matrix from proximity matrix}
\usage{
affinity_matrix(S, N = 10)
}
\arguments{
\item{S}{square matrix}

\item{N}{nearest neighbours}
}
\description{
Extract affinity matrix from proximity matrix
}
