% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder}
\alias{autoencoder}
\title{Learn a deep autoencoder}
\usage{
autoencoder(x, seed = 1234)
}
\arguments{
\item{x}{Numeric expression matrix}

\item{seed}{Random seed integer}
}
\description{
Learn a deep autoencoder
}
