% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_affinity_matrix.R
\name{compute_affinity_matrix}
\alias{compute_affinity_matrix}
\title{Generate an affinity matrix for a single run}
\usage{
compute_affinity_matrix(X_train, X_test, num_trees, N, ..., verbose = TRUE,
  seed = 1234)
}
\arguments{
\item{X_train}{training set}

\item{X_test}{testing set}

\item{num_trees}{number of trees}

\item{N}{number of nearest neighbours for affinity matrix}

\item{...}{additional parameters to ranger}

\item{verbose}{control verbosity}
}
\description{
Generate an affinity matrix for a single run
}
