% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.R
\name{cytorf}
\alias{cytorf}
\title{Perform unsupervised clusering using random forest}
\usage{
cytorf(X, channels = NULL, num_trees = 2000, delta = 0.1, seed = 1234,
  verbose = FALSE, ...)
}
\arguments{
\item{X}{a data matrix for the flow cytometry data, it needs to have
at least two columns.}

\item{channels}{a vector of two channel names or their corresponding indices.
       When it is left unspecified, all the variables will be
included in clustering.}

\item{num_trees}{number of trees}

\item{seed}{random seed that controls clustering reproducibility}

\item{verbose}{boolean level of verbosity (default: FALSE)}

\item{...}{additional parameters to be passed to ranger function}

\item{N}{number of neighbours for calculation of affinity matrix.}
}
\description{
Perform unsupervised clusering using random forest
}
