% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepflow.R
\name{deepflow}
\alias{deepflow}
\title{DeepFlow structure detection algorithm}
\usage{
deepflow(x = NULL, transform = FALSE, gui = TRUE, seed = 1234)
}
\arguments{
\item{x}{numerical matrix}

\item{transform}{boolean value indicating whether arcsinh/5 transformation
should take place}

\item{gui}{boolean value indicating whether a shiny application should be launched}

\item{seed}{random seed to ensure reproducibility}
}
\description{
DeepFlow structure detection algorithm
}
