% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepflow_init.R
\name{deepflow_init}
\alias{deepflow_init}
\title{Initiliase deepflow application}
\usage{
deepflow_init(dbname, dbpath)
}
\arguments{
\item{dbname}{database name to be used across the application}

\item{dbpath}{path pointing to location of SQLite database on disk}
}
\description{
Initiliase deepflow application
}
