% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.R
\name{predict.cytorf}
\alias{predict.cytorf}
\title{Predict cytorf clusters from a supervised model}
\usage{
\method{predict}{cytorf}(model, newdata, verbose = FALSE)
}
\arguments{
\item{model}{cytorf model}

\item{newdata}{data.frame object}

\item{verbose}{boolean control of verbosity levels. Default is FALSE.}
}
\description{
Predict cytorf clusters from a supervised model
}
