% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.R
\name{cytorf}
\alias{cytorf}
\title{Perform unsupervised clusering using random forest}
\usage{
cytorf(X, Y = NULL, channels = NULL, num_trees = 150, N = 5,
  subsample = 0.25, seed = 1234, verbose = FALSE, ...)
}
\arguments{
\item{X}{a data matrix for the flow cytometry data, it needs to have
at least two columns.}

\item{channels}{a vector of two channel names or their corresponding indices.
       When it is left unspecified, all the variables will be
included in clustering.}

\item{num_trees}{number of trees}

\item{N}{number of neighbours for calculation of affinity matrix.}

\item{subsample}{double indicating a fraction of random observations to
chose for calculation of proximity matrix. Range is 0-1.
This parameter can be used to speed up model construction#'                    by sampling at random a small partition of the dataset.}

\item{seed}{random seed that controls clustering reproducibility}

\item{verbose}{boolean level of verbosity (default: FALSE)}

\item{...}{additional parameters to be passed to ranger function}
}
\description{
Perform unsupervised clusering using random forest
}
